# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Visitor Management System API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/checkInOut": {
      "get": {
        "summary": "Get all check-in/out records",
        "tags": [
          "CheckInOut"
        ],
        "responses": {
          "200": {
            "description": "List of check-in/out records"
          }
        }
      },
      "post": {
        "summary": "Create a new check-in/out record",
        "tags": [
          "CheckInOut"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "visitorId": {
                    "type": "integer"
                  },
                  "location": {
                    "type": "string"
                  },
                  "checkInTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "checkOutTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Record created"
          }
        }
      }
    },
    "/api/v1/checkInOut/{checkin_id}": {
      "get": {
        "summary": "Get a check-in/out record by ID",
        "tags": [
          "CheckInOut"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "checkin_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the check-in/out record"
          }
        ],
        "responses": {
          "200": {
            "description": "Record found"
          },
          "404": {
            "description": "Record not found"
          }
        }
      },
      "put": {
        "summary": "Update a check-in/out record",
        "tags": [
          "CheckInOut"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "checkin_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the record to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "visitorId": {
                    "type": "integer"
                  },
                  "location": {
                    "type": "string"
                  },
                  "checkInTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "checkOutTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Record updated"
          },
          "404": {
            "description": "Record not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a check-in/out record",
        "tags": [
          "CheckInOut"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "checkin_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the record to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Record deleted"
          },
          "404": {
            "description": "Record not found"
          }
        }
      }
    },
    "/api/v1/checkInOut/host/{user_id}": {
      "get": {
        "summary": "Get all visitors for a specific host",
        "tags": [
          "CheckInOut"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Host user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of visitors for the host"
          }
        }
      }
    },
...
    "/api/v1/checkInOut/checkin": {
      "post": {
        "summary": "Check in a visitor by RFID",
        "tags": [
          "CheckInOut"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rfidToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visitor checked in"
          }
        }
      }
    },
    "/api/v1/checkInOut/checkout": {
      "post": {
        "summary": "Check out a visitor by RFID",
        "tags": [
          "CheckInOut"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rfidToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visitor checked out"
          }
        }
      }
    },
    "/api/v1/checkInOut/end-meeting": {
      "post": {
        "summary": "End a meeting manually",
        "tags": [
          "CheckInOut"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "checkinId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Meeting ended"
          }
        }
      }
    },
    "/api/v1/checkInOut/extend-stay": {
      "post": {
        "summary": "Extend a visitor's stay",
        "tags": [
          "CheckInOut"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "checkinId": {
                    "type": "integer"
                  },
                  "newCheckoutTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stay extended"
          }
        }
      }
    },
    "/api/v1/checkin/checkin": {
      "post": {
        "summary": "Check in a visitor",
        "tags": [
          "Checkin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "appointmentId"
                ],
                "properties": {
                  "appointmentId": {
                    "type": "integer",
                    "description": "The appointment ID of the visitor"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visitor successfully checked in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "checkin": {
                      "type": "object",
                      "description": "Check-in record details"
                    },
                    "token": {
                      "type": "string",
                      "description": "Token generated for visitor"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or error message"
          }
        }
      }
    },
    "/api/v1/checkin/checkout": {
      "post": {
        "summary": "Check out a visitor",
        "tags": [
          "Checkin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tokenId"
                ],
                "properties": {
                  "tokenId": {
                    "type": "string",
                    "description": "Token ID assigned to the visitor at check-in"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visitor successfully checked out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "checkout": {
                      "type": "object",
                      "description": "Check-out record details"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or error message"
          }
        }
      }
    },
...
    "/api/v1/user/reception/checkin": {
      "post": {
        "summary": "Check-in visitor (admin or receptionist)",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Check-in successful"
          }
        }
      }
    },
...
  },
...
}